requirements.txt
flask
azure-cosmos
uuid
python-dotenv